+++ CHANGELOG.md
@@ -4,10 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# Unreleased
+# 2.2.1 2021-07-16
 
 ### Added
-- added `removeDatabase` function.
+- added `removeDatabase` method to `VuexOffline` class.
 
 # 2.2.0 2021-07-16